Error table posts

+++ database/migrations/2021_03_03_152854_create_posts_table.php
@@ -14,8 +14,8 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id'); //Author
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id'); //Author
             $table->string('title')->unique();
             $table->text('content');
             $table->timestamps();
